sshprocess.cpp: don't use QProcess::start (QString).
Qt is trying to be too smart and causes big trouble. Instead, use
QProcess::start (QString, QStringList) and pass the arguments as a list.

On Windows, Qt will automatically double quote the arguments and
duplicate escaped double quotes or escape non-escaped double quotes.

On UNIX-like platforms, each element of the list is passed as a unique
argv element, so there's no need for quoting them (that's only a
shell-internal thing to group arguments.)