src/sshmasterconnection.cpp: turns out libssh < 0.6.0 doesn't support fetching the host, port and username parameters at all.
We're in the clear for non-SSH-proxied connections and can use config
file parsing, but have to disable config file parsing in the proxy
settings, as there's no way to query the remote endpoint information. As
a result we will be unable to open a new socket to the remote endpoint
and proxying will fail badly.