src/{httpbrokerclient,onmainwindow,sshmasterconnection}.{cpp,h}: introduce new passphrase_type enum and replace the old verificationCode boolean value with that.
Allows using more messages, including a new one for plain password
prompts.

Adapt usage accordingly.

The implementation is ugly, but there's no good way around that (short
of using C++11 features). The enum would really logically belong to
ONMainWindow, but since we also have to use it in SshMasterConnection,
that may not include onmainwindow.h to avoid a circular dependency, it
has to be part of SshMasterConnection for now.